cleanup code

+++ src/main/java/cn/academy/support/rf/TileRFInput.java
@@ -27,7 +27,7 @@ public class TileRFInput extends TileGeneratorBase implements IEnergyReceiver
     public int receiveEnergy(EnumFacing from, int maxReceive,
             boolean simulate) {
         int amount = (int) rf2if(maxReceive);
-        return (int) (maxReceive - if2rf(addEnergy(amount, simulate)));
+        return maxReceive - if2rf(addEnergy(amount, simulate));
     }
 
     @Override
@@ -37,7 +37,7 @@ public class TileRFInput extends TileGeneratorBase implements IEnergyReceiver
 
     @Override
     public int getMaxEnergyStored(EnumFacing from) {
-        return (int) if2rf(2000);
+        return if2rf(2000);
     }
 
     @Override
